high --> lower priority node creation/handshake
Start nodes with high priority (both for the parent and the child) until the handshake is over. The parent, then, remains high priority if and only if it has other children currently at high priority; the child always switches to the appropriate priority as understood by the parent's pre-creating-node priority.